Add warning for possible incomplete scan-summaries
In 2020.08 and 2020.10 the codelocations/{}/scan-summaries endpoint only
returns 1 entry even though more may exist. See HUB-25835.

This bug makes it nearly impossible to determine if too frequent scanning to a
given codelocation is occurring.  It is possible to look at the journal events
for a project version but this method is both slow and when codelocations are
mapped/unmapped it is unreliable.

I searched for possible work-arounds in the REST API but did not find any
satisfactory ones and thus the best we can do is provide a warning about the
issue.